+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed crash on forge
+- Updated for Tardis Refined Breaking changes in 2.0.6